Add new tests, and fix subsequent issues

+++ solutions.digamma.damas.rs/src/main/java/solutions/digamma/damas/rs/content/FolderSerialization.java
@@ -37,7 +37,7 @@ public class FolderSerialization extends FileSerialization implements Folder {
     }
 
     @Override
-    public void expandContent(long depth) {
+    public void expandContent(int depth) {
     }
 
     @Override
